refactor: 모듈화 및 공지사항 캐러셀을 위한 파일 이름, 폴더 구조 변경
