+++ src/features/main/components/cards/EventCard.styles.ts
