+++ src/features/main/components/cards/EventCard.types.ts
