+++ src/features/main/components/cards/Progress.styles.ts
