+++ src/features/main/components/cards/index.ts
@@ -1,0 +1,1 @@
+export { default as EventCard } from './EventCard';
