+++ src/features/main/components/carousels/EventCarousels.styles.ts
