+++ src/features/main/components/carousels/index.ts
@@ -1,0 +1,1 @@
+export { default as EventCarousels } from './EventCarousels';
